+++ src/components/AddEventToPublishedScheduleForm/AddEventToPublishedScheduleForm.jsx
@@ -5,15 +5,21 @@ import {
   Input,
   FormControl,
   FormErrorMessage,
+  Button,
   Textarea,
   Checkbox,
+  useToast,
   Heading,
   Flex,
   Text,
+  HStack,
+  useDisclosure,
+  Spacer
 } from '@chakra-ui/react';
 import { yupResolver } from '@hookform/resolvers/yup';
 import { useForm } from 'react-hook-form';
 import { useContext, useEffect, useState } from 'react';
+import { NPOBackend } from '../../utils/auth_utils';
 import * as yup from 'yup';
 import {
   seasonOptions,
@@ -21,10 +27,12 @@ import {
   subjectOptions,
   eventOptions,
 } from '../Catalog/SearchFilter/filterOptions';
+import useSearchFilters from '../Catalog/SearchFilter/useSearchFilters';
 import Dropdown from '../Dropdown/Dropdown';
 import PropTypes from 'prop-types';
 import { PlannerContext } from '../Planner/PlannerContext';
 import PlannedEvent, { convertTimeToMinutes } from '../Planner/PlannedEvent';
+import RemoveTimelineEventModal from '../Planner/RemoveTimelineEventModal';
 
 const schema = yup.object({
     startTime: yup.string().required('Start time is required'),
@@ -44,15 +52,16 @@ const schema = yup.object({
     tentative: yup.boolean()
 });
 
-const AddEventToPublishedScheduleForm = ({ submitData, filterObj }) => {
-  const { plannedEventsContext, editContext, currEventContext } = useContext(PlannerContext);
+const AddEventToPublishedScheduleForm = ({ closeForm }) => {
+  const { plannedEventsContext, dayId, editContext, currEventContext } = useContext(PlannerContext);
   const [plannedEvents, setPlannedEvents] = plannedEventsContext;
-  const eventData = currEventContext;
-  const isEdit = editContext;
-  const { filters, filterValues } = filterObj;
+  const [eventData, setCurrEvent] = currEventContext;
+  const [isEdit, setIsEdit] = editContext;
+  const { filters, filterValues } = useSearchFilters();
   const [seasonFilter, yearFilter, subjectFilter, eventFilter] = filters;
   const [checkboxVal, setCheckboxVal] = useState(undefined);
   const [formData, setFormData] = useState({...eventData});
+  const { isOpen: isRemoveOpen, onOpen: onRemoveOpen, onClose: onRemoveClose } = useDisclosure();
 
   useEffect(() => {
     if (Object.keys(eventData).length === 0) {
@@ -99,6 +108,7 @@ const AddEventToPublishedScheduleForm = ({ submitData, filterObj }) => {
   // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [formData])
 
+  const toast = useToast();
   const {
     setValue,
     register,
@@ -108,6 +118,153 @@ const AddEventToPublishedScheduleForm = ({ submitData, filterObj }) => {
   } = useForm({
     resolver: yupResolver(schema),
   });
+
+  const handleCancel = () => {
+    if (isEdit) {
+      // let reAddedEvent = plannedEvents.filter(e => e.id == -1)[0];
+      // reAddedEvent.id = eventData.id;
+      const reAddedEvent = new PlannedEvent(
+        eventData.id,
+        eventData.title,
+        convertTimeToMinutes(eventData.startTime),
+        convertTimeToMinutes(eventData.endTime),
+        eventData.host,
+        !eventData.confirmed
+      )
+      setPlannedEvents([...plannedEvents.filter(e => e.id != -1), reAddedEvent]);
+    } else {
+      setPlannedEvents(plannedEvents.filter(e => e.id != -1));
+    }
+    setCurrEvent({});
+    setIsEdit(false);
+    closeForm();
+  }
+
+  const currentCatalogDataHasChanged = (originalData, currData) => {
+    // keys: title,host,description,eventType,subject,year,season
+    for (let key of Object.keys(currData)) {
+      if (originalData[key] === undefined || originalData[key] !== currData[key]) {
+        return true;
+      }
+    }
+    return false;
+  }
+
+  const displayToast = () => {
+    let toastTitle = 'Success!';
+    let toastDescription = 'Added event to day.';
+    if (isEdit) {
+      toastTitle = 'Saved!';
+      toastDescription = 'Changes to event were saved.';
+    }
+
+    toast({
+      title: toastTitle,
+      description: toastDescription,
+      status: 'success',
+      variant: 'subtle',
+      position: 'top-right',
+      containerStyle: {
+        mt: '6rem',
+      },
+      duration: 3000,
+      isClosable: true,
+    });
+  }
+
+  const submitData = async (data) => {
+    try {
+      // eslint-disable-next-line no-unused-vars
+      const { title, host, description, tentative, startTime, endTime } = data;
+      const season = filterValues.season;
+      const eventType = filterValues.eventType;
+      const year = filterValues.year;
+      const subject = filterValues.subject;
+      console.log('submmitted data', data, season, eventType, year, subject);
+
+      toast.closeAll();
+
+      const catalogDataChanged = currentCatalogDataHasChanged(eventData, {
+        title,
+        host,
+        description,
+        eventType,
+        subject,
+        year,
+        season
+      });
+
+      let catalogEventId = eventData.id;
+      if(isEdit) {
+        catalogEventId = eventData.eventId;
+      }
+
+      // not editing timeline event AND (changed catalog data OR is completely new event)
+      if (!isEdit && (catalogDataChanged || !catalogEventId)) {
+        const catalogResponse = await NPOBackend.post(`/catalog`, {
+          title,
+          host,
+          description,
+          eventType,
+          subject,
+          year,
+          season
+        });
+
+        catalogEventId = catalogResponse.data.id;
+      }
+
+      let publishedScheduleReponse;
+      let plannedEventId;
+      if (isEdit) {
+        if (catalogDataChanged) {
+          const updateCatalog = await NPOBackend.put(`/catalog/${catalogEventId}`, {
+            title, host, description, eventType, subject, year, season
+          });
+          catalogEventId = updateCatalog.data[0].id;
+        }
+        // Send a PUT request
+        publishedScheduleReponse = await NPOBackend.put(`/published-schedule/${eventData.id}`, {
+          eventId: catalogEventId,
+          confirmed: !checkboxVal,
+          startTime,
+          endTime,
+          cohort: year,
+        });
+        plannedEventId = eventData.id;
+      } else {
+        // Send a POST request to the appropriate backend route
+        publishedScheduleReponse = await NPOBackend.post('/published-schedule', {
+          eventId: catalogEventId,
+          dayId,
+          confirmed: !checkboxVal,
+          startTime,
+          endTime,
+          cohort: year,
+        });
+        plannedEventId = publishedScheduleReponse.data.id;
+      }
+      const timelineEventsWithoutCurrent = plannedEvents.filter(e => (e.id != -1 && e.id != eventData.id));
+      const newPlannedEvent = new PlannedEvent(
+        plannedEventId,
+        title,
+        convertTimeToMinutes(startTime),
+        convertTimeToMinutes(endTime),
+        host,
+        checkboxVal
+      );
+      setPlannedEvents([...timelineEventsWithoutCurrent, newPlannedEvent]);
+      setFormData({tentative: false});
+
+      reset();
+      setCurrEvent({});
+
+      displayToast();
+      closeForm();
+    } catch (error) {
+      console.log(error);
+    }
+  };
 
   return (
     <Box borderRadius="5px">
@@ -277,24 +434,22 @@ const AddEventToPublishedScheduleForm = ({ submitData, filterObj }) => {
               </Box>
             </Box>
         </Box>
-        {/* <Stack spacing={2} justifyContent="right" direction="row" pb="1.5rem" mt="0.5rem">
-          {isEdit &&
-            <>
-              <Button mt="1rem" variant='outline' onClick={onRemoveOpen}>Delete</Button>
-              <Spacer />
-            </>
-          }
-          <Button htype="submit" mt="1rem" mr="1rem" onClick={handleCancel}>Cancel</Button>
-          <Button colorScheme="blue" type="submit" mt="1rem">{isEdit ? 'Save' : 'Add Event'}</Button>
-        </Stack> */}
       </form>
+      <HStack spacing={2} pb="1.5rem" mt="0.5rem">
+        {isEdit &&
+          <Button variant="outline" mt="1rem" mr="1rem" onClick={onRemoveOpen}>Delete</Button>
+        }
+        <Spacer />
+        <Button htype="submit" mt="1rem" mr="1rem" onClick={handleCancel}>Cancel</Button>
+        <Button form="add-ps-event-form" colorScheme="blue" type="submit" mt="1rem">{isEdit ? 'Save' : 'Add Event'}</Button>
+      </HStack>
+      <RemoveTimelineEventModal onClose={onRemoveClose} isOpen={isRemoveOpen} deleteItemId={eventData.id ? eventData.id : -1} closeForm={closeForm}/>
     </Box>
   );
 };
 
 AddEventToPublishedScheduleForm.propTypes = {
-  submitData: PropTypes.func,
-  filterObj: PropTypes.object,
+  closeForm: PropTypes.func
 };
 
 export default AddEventToPublishedScheduleForm;
